Move kak_lsp_custom.rs into document_symbol.rs
No need for extra file
